+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `refundFromTransactionFeeCurrencyIsoCode` field to localPayment
 * Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
 * Add `threeDSecureVersion`, `authenticationResponse`, `directoryResponse`, `cavvAlgorithm` and `dsTransactionId` to 3DS pass thru fields
+* Add `payerInfo` field to paymentMethodNonce
 
 ## 2.17.0
 * Update @braintree/wrap-promise to v2.0.0