+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add `fireSafe` support on transaction sale
 * Add `propertyPhone` support on transaction sale
 * Add `additionalCharges` support on transaction sale
+* Add `PostalCodeIsRequiredForCardBrandAndProcessor` to validation errors
 
 ## 2.17.0
 * Update @braintree/wrap-promise to v2.0.0